Updating to GDX version 1.14.0

+++ gdx-graph-shader-util/src/com/gempukku/libgdx/graph/util/model/GeometryMeshCapture.java
@@ -193,7 +193,7 @@ public class GeometryMeshCapture implements MeshPartBuilder {
     }
 
     @Override
-    public short lastIndex() {
+    public int lastIndex() {
         return (short) (nextVertexIndex - 1);
     }
